Volume and Punk additions.
- Made volume sliders work.
- Added 3rd Phase to the Punk.
- Added some missing animations and sounds to the Punk.
- Added more NextPattern conditions to the Bosses (intended to be used with the Punk).
- Changes to some Punk attacks.
- Fixed dash not bouncing the correct amount of times.
- Fixed Dash not correctly waiting for charge time.
- Updated Punk music and imported Xing music.
- Fixed Sound volume and pitch using int instead of float.
- Added the ability to stop sounds when the Boss die or win.
- Made the player unable to take damage after defeating a Boss.

+++ ComicSans/Assets/Scripts/Audio/Sound.cs
@@ -14,8 +14,8 @@ public class Sound
 	// For the FX type a random clip from the array will be played each time.
 	public AudioClip[] clips = new AudioClip[]{};
 
-	public int volume = 1;
-	public int pitch = 1;
+	public float volume = 1.0f;
+	public float pitch = 1.0f;
 
 	public int maxSimultaneousSources = 3;
